Move the admin UI off /admin to /flightdeck
The whole internet probes /admin continuously. Nothing there authenticates
any more: the route is a 404 with no page behind it, and the admin UI now
lives at /flightdeck with the same HTTP Basic gate in front of it.

This is noise reduction rather than a security control, and the difference
matters because the name is visible in a public repository. The threat being
addressed is the untargeted scanner working through a wordlist, not someone
who has decided to attack this project specifically. Basic auth remains the
thing actually standing in the way. The name deliberately avoids cockpit,
backstage, adminer and portainer, which are themselves probed routinely.

The risky part was not the directory move but the twenty-three absolute paths
hard-coded across the index, the four section pages, and the ADMIN_PATH
constants that revalidatePath uses in both actions.ts files. Missing one gives
either a dead link or a silently stale page after a mutation, and neither the
type checker nor the tests would catch it, since nothing in the suite asserts
a URL. Verification was done against a running production build instead:
/admin and /admin/feeds return 404, /flightdeck and /flightdeck/feeds return
401 carrying the Fetchlinks admin realm, correct credentials return 200, wrong
ones 401, and the public listing is untouched.

The FETCHLINKS_ADMIN_* variable names, the Fetchlinks admin realm and the
admin-* CSS classes are unchanged on purpose. They are internal names rather
than URLs, and renaming them would break the environment variable contract
with Vercel for no gain. The example admin username is no longer "admin",
which was teaching exactly the wrong default.

Note for anyone repeating this: next build regenerates .next/dev/types/
validator.ts but does not prune routes that no longer exist, so the first
build after a rename fails with TS2307 errors naming files that are correctly
absent. Delete .next.
