+++ web/.env.production.example
@@ -6,6 +6,6 @@ DATABASE_URL=postgresql://fetchlinks_web:__PASSWORD__@__NEON_POOLED_HOST__/fetch
 NODE_ENV=production
 NEXT_TELEMETRY_DISABLED=1
 
-# Credentials for the /admin/* routes (HTTP Basic). Leave unset to disable.
-FETCHLINKS_ADMIN_USER=admin
+# Credentials for the /flightdeck/* routes (HTTP Basic). Leave unset to disable.
+FETCHLINKS_ADMIN_USER=change-me
 FETCHLINKS_ADMIN_PASS=change-me
